refactor($renderer): mv initialize.js to usecases folder
