+++ src/renderer/main.js
@@ -7,7 +7,7 @@ import store from './store';
 import Vuetify from 'vuetify';
 import 'vuetify/dist/vuetify.min.css'; // Ensure you are using css-loader
 import theme from '../constants/theme';
-import initialize from './initialize';
+import initialize from './usecases/initialize';
 
 if (!process.env.IS_WEB) Vue.use(require('vue-electron'));
 Vue.config.productionTip = false;
